Schema validation fix
Former-commit-id: 594421bdda3cf6701a1878d8422342bcbd7c2b18
Former-commit-id: d7efb39bfc894bb7d9ffec6ca29e79c46d88cd46

+++ kernel/kernel-masterdata-service/src/test/java/io/mosip/kernel/masterdata/test/service/LocationHierarchyTest.java
@@ -39,7 +39,7 @@ public class LocationHierarchyTest {
 		Location locationHierarchy = new Location();
 		locationHierarchy.setCode("IND");
 		locationHierarchy.setName("INDIA");
-		locationHierarchy.setHierarchyLevel(0);
+		locationHierarchy.setHierarchyLevel((short)0);
 		locationHierarchy.setHierarchyName(null);
 		locationHierarchy.setParentLocCode(null);
 		locationHierarchy.setLangCode("HIN");
@@ -50,7 +50,7 @@ public class LocationHierarchyTest {
 		Location locationHierarchy1 = new Location();
 		locationHierarchy1.setCode("KAR");
 		locationHierarchy1.setName("KARNATAKA");
-		locationHierarchy1.setHierarchyLevel(1);
+		locationHierarchy1.setHierarchyLevel((short)1);
 		locationHierarchy1.setHierarchyName(null);
 		locationHierarchy1.setParentLocCode("TEST");
 		locationHierarchy1.setLangCode("KAN");
